Restructure the view package. #18
