+++ app-paper/src/main/java/com/paper/editor/view/canvas/IPaperContext.kt
@@ -18,20 +18,36 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-package com.paper.editor.view
+package com.paper.editor.view.canvas
 
-import android.graphics.Bitmap
-import com.paper.editor.data.DrawViewPortEvent
-import com.paper.editor.widget.IPaperWidget
-import io.reactivex.Observable
+import android.view.ViewConfiguration
 
-interface IPaperWidgetView {
+/**
+ * The context gives the global editor settings and the ability to map point
+ * from one coordinate to other coordinate.
+ */
+interface IPaperContext {
 
-    fun bindWidget(widget: IPaperWidget)
+    // Rendering //////////////////////////////////////////////////////////////
 
-    fun unbindWidget()
+    fun getMinStrokeWidth(): Float
 
-    fun takeSnapshot(): Bitmap
+    fun getMaxStrokeWidth(): Float
 
-    fun onDrawViewPort(): Observable<DrawViewPortEvent>
+    /**
+     * Map the point observed in the Model world to the View world.
+     */
+    fun mapM2V(x: Float, y: Float): FloatArray
+
+    // Gesture ////////////////////////////////////////////////////////////////
+
+    fun getViewConfiguration(): ViewConfiguration
+
+    fun getTouchSlop(): Float
+
+    fun getTapSlop(): Float
+
+    fun getMinFlingVec(): Float
+
+    fun getMaxFlingVec(): Float
 }
