+++ app-paper/src/main/java/com/paper/editor/view/canvas/ScrapWidgetView.kt
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-package com.paper.editor.view
+package com.paper.editor.view.canvas
 
 import android.graphics.*
 import android.os.Looper
